Move GB18030 tests to feature test projects
Move each GB18030 compliance test into the test project for the feature it exercises. Share the common helper and conformance resource through a projitems import so feature projects retain the same test data without duplication.

This lets existing feature-project platform exclusions apply naturally and avoids bringing the regex test project's Roslyn dependencies into NativeAOT mobile compliance runs.
