+++ src/libraries/Common/Common.Tests.slnx
@@ -6,7 +6,6 @@
   </Folder>
   <Folder Name="/tests/">
     <Project Path="tests/Common.Tests.csproj" />
-    <Project Path="tests/ComplianceTests/Compliance.Tests.csproj" />
     <Project Path="tests/StreamConformanceTests/StreamConformanceTests.csproj" />
     <Project Path="tests/TestUtilities/TestUtilities.csproj" />
   </Folder>
