+++ src/libraries/System.Runtime/tests/System.IO.Tests/GB18030/DirectoryTestBase.cs
