+++ src/libraries/System.Runtime/tests/System.IO.Tests/GB18030/FileTests.cs
