+++ src/libraries/System.Text.Encoding.CodePages/tests/GB18030/EncodingTests.cs
