+++ src/libraries/System.Text.RegularExpressions/tests/FunctionalTests/GB18030/RegexTests.cs
